+++ .idea/misc.xml
@@ -3,6 +3,10 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/kuliah/Bangkit/fundamental android/OCR_C22-KY03/app/src/main/res/layout/activity_ktp.xml" value="0.1762917933130699" />
+        <entry key="..\:/kuliah/Bangkit/fundamental android/OCR_C22-KY03/app/src/main/res/layout/activity_ktp_result.xml" value="0.1762917933130699" />
+        <entry key="..\:/kuliah/Bangkit/fundamental android/OCR_C22-KY03/app/src/main/res/layout/activity_scan.xml" value="0.33" />
+        <entry key="..\:/kuliah/Bangkit/fundamental android/OCR_C22-KY03/app/src/main/res/layout/item_row_history.xml" value="0.1762917933130699" />
         <entry key="app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.215" />
         <entry key="app/src/main/res/drawable/border_button.xml" value="0.142" />
         <entry key="app/src/main/res/drawable/ic_baseline_insert_photo_24.xml" value="0.1565" />
@@ -10,7 +14,7 @@
         <entry key="app/src/main/res/drawable/ic_launcher_background.xml" value="0.215" />
         <entry key="app/src/main/res/drawable/rounded_button.xml" value="0.142" />
         <entry key="app/src/main/res/layout/activity_detail.xml" value="0.28229166666666666" />
-        <entry key="app/src/main/res/layout/activity_form.xml" value="0.260545854598007" />
+        <entry key="app/src/main/res/layout/activity_form.xml" value="0.2682291666666667" />
         <entry key="app/src/main/res/layout/activity_history.xml" value="0.240625" />
         <entry key="app/src/main/res/layout/activity_ktp.xml" value="0.33" />
         <entry key="app/src/main/res/layout/activity_ktp_result.xml" value="0.26510416666666664" />
