+++ app/build.gradle
@@ -54,7 +54,7 @@ dependencies {
 
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation "com.squareup.retrofit2:converter-gson:2.9.0"
-    implementation "com.squareup.okhttp3:logging-interceptor:5.0.0-alpha.5"
+    implementation "com.squareup.okhttp3:logging-interceptor:5.0.0-alpha.6"
 
     implementation "androidx.lifecycle:lifecycle-reactivestreams-ktx:2.4.1"
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.4.1"
